fix: address Copilot review comments on fixed-point overflow PR
- Use 1ULL shifts in Block::setFixed to avoid UB at 64-bit widths
- Fix UFixed.maxValue() to return the signed-range max that C++ actually
  supports (the getFixed/setFixed path is always two's-complement, so
  values above 2^(N-1)-1 are not representable for UFixed)
- Update UFixed.maximum test assertion from 4095.9375 to 2047.9375
- Replace try/except with pytest.raises(Exception, match=...) for
  stricter, idiom-consistent overflow tests

+++ src/rogue/interfaces/memory/Block.cpp
@@ -2012,9 +2012,9 @@ void rim::Block::setFixed(const double& val, rim::Variable* var, int32_t index) 
     // Convert
     int64_t fPoint = static_cast<int64_t>(round(val * pow(2, var->binPoint_)));
 
-    // Compute representable range in integer domain
-    int64_t maxInt = (1LL << (var->valueBits_ - 1)) - 1;
-    int64_t minInt = -(1LL << (var->valueBits_ - 1));
+    // Compute representable range in integer domain (use 1ULL to avoid UB for 64-bit widths)
+    int64_t maxInt = static_cast<int64_t>((1ULL << (var->valueBits_ - 1)) - 1);
+    int64_t minInt = -static_cast<int64_t>(1ULL << (var->valueBits_ - 1));
 
     // Check for overflow (rounding may push value beyond representable range)
     if (fPoint > maxInt || fPoint < minInt)
